+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 3.0.0
 * BREAKING: a bit of renaming & export cleaning.
+* New `FetcherConfig.silent` config for cases where loader & error should not be displayed.
 
 ## 2.0.1
 * BREAKING: rename `FetcherConfigErrorData` to `FetchErrorData`.